1、首页界面修改 2、登录接口返回状态修改 3、退出登录增加返回值 4、系统参数以及操作日志界面修改

+++ src/components/header.jsx
@@ -31,10 +31,14 @@ class DHeader extends Component {
     quit = _ => {
         
         logout().then(res=>{
-            // 清除缓存
-            localStorage.clear();
-            this.props.reset()
-            hashHistory.replace('/login')
+            if (res.success == 1) {
+                // 清除缓存
+                localStorage.clear();
+                this.props.reset()
+                hashHistory.replace('/login')
+            }else {
+                message.error(res.message)
+            }
         })
     }
